Web App that asks the user to enter user name and password and does a certain action based on credentials provided.
Web App that asks the user to enter user name and password. The user name can be anything. The password is set to  "abc." If the user enters the password correctly, the user is directed to the welcome page and the message "you're in," is displayed. If the user enters the password incorrectly, the user is still directed to the welcome page, but this time, the message "try again," is displayed.